Add config and handlers to inbound connector base
